Adding an article to week 02 of articles of the week

+++ myadmin/views.py
@@ -13,7 +13,7 @@ from django.views import generic
 from django.views.decorators.http import require_http_methods
 from django.views.generic import TemplateView
 
-from comment.models import Comment
+from comment.models import Comment, Danmaku, Reply
 from helpers import get_page_list, AdminUserRequiredMixin, ajax_required, SuperUserRequiredMixin, send_html_email
 from users.models import User, Feedback
 from video.models import Video, Classification
@@ -235,6 +235,48 @@ class CommentListView(AdminUserRequiredMixin, generic.ListView):
         return Comment.objects.filter(content__contains=self.q).order_by('-timestamp')
 
 
+class DanmakuListView(AdminUserRequiredMixin, generic.ListView):
+    model = Danmaku
+    template_name = 'myadmin/danmaku_list.html'
+    context_object_name = 'danmaku_list'
+    paginate_by = 10
+    q = ''
+
+    def get_context_data(self, *, object_list=None, **kwargs):
+        context = super(DanmakuListView, self).get_context_data(**kwargs)
+        paginator = context.get('paginator')
+        page = context.get('page_obj')
+        page_list = get_page_list(paginator, page)
+        context['page_list'] = page_list
+        context['q'] = self.q
+        return context
+
+    def get_queryset(self):
+        self.q = self.request.GET.get("q", "")
+        return Danmaku.objects.filter(content__contains=self.q).order_by('-timestamp')
+
+
+class ReplyListView(AdminUserRequiredMixin, generic.ListView):
+    model = Reply
+    template_name = 'myadmin/reply_list.html'
+    context_object_name = 'reply_list'
+    paginate_by = 10
+    q = ''
+
+    def get_context_data(self, *, object_list=None, **kwargs):
+        context = super(ReplyListView, self).get_context_data(**kwargs)
+        paginator = context.get('paginator')
+        page = context.get('page_obj')
+        page_list = get_page_list(paginator, page)
+        context['page_list'] = page_list
+        context['q'] = self.q
+        return context
+
+    def get_queryset(self):
+        self.q = self.request.GET.get("q", "")
+        return Reply.objects.filter(content__contains=self.q).order_by('-timestamp')
+
+
 @ajax_required
 @require_http_methods(["POST"])
 def comment_delete(request):
@@ -242,6 +284,28 @@ def comment_delete(request):
         return JsonResponse({"code": 1, "msg": "无删除权限"})
     comment_id = request.POST['comment_id']
     instance = Comment.objects.get(id=comment_id)
+    instance.delete()
+    return JsonResponse({"code": 0, "msg": "success"})
+
+
+@ajax_required
+@require_http_methods(["POST"])
+def danmaku_delete(request):
+    if not request.user.is_superuser:
+        return JsonResponse({"code": 1, "msg": "无删除权限"})
+    danmaku_id = request.POST['danmaku_id']
+    instance = Danmaku.objects.get(id=danmaku_id)
+    instance.delete()
+    return JsonResponse({"code": 0, "msg": "success"})
+
+
+@ajax_required
+@require_http_methods(["POST"])
+def reply_delete(request):
+    if not request.user.is_superuser:
+        return JsonResponse({"code": 1, "msg": "无删除权限"})
+    reply_id = request.POST['reply_id']
+    instance = Reply.objects.get(id=reply_id)
     instance.delete()
     return JsonResponse({"code": 0, "msg": "success"})
